refactor: more consistent js-api docs (#1463)
