+++ tooling/api/src/fs.ts
@@ -47,20 +47,21 @@ export interface FsBinaryFileOption {
 
 export interface FileEntry {
   path: string
-  // name of the directory/file
-  // can be null if the path terminates with `..`
+  /**
+   * Name of the directory/file
+   * can be null if the path terminates with `..`
+   */
   name?: string
-  // children of this entry if it's a directory; null otherwise
+  /** Children of this entry if it's a directory; null otherwise */
   children?: FileEntry[]
 }
 
 /**
- * @name readTextFile
- * @description Reads a file as text
- * @param {string} filePath path to the file
- * @param {FsOptions} [options] configuration object
- * @param {BaseDirectory} [options.dir] base directory
- * @return {Promise<string>}
+ * Reads a file as text.
+ *
+ * @param filePath Path to the file
+ * @param [options]
+ * @returns A promise resolving to a string of the file content.
  */
 async function readTextFile(
   filePath: string,
@@ -77,12 +78,11 @@ async function readTextFile(
 }
 
 /**
- * @name readBinaryFile
- * @description Reads a file as binary
- * @param {string} filePath path to the file
- * @param {FsOptions} [options] configuration object
- * @param {BaseDirectory} [options.dir] base directory
- * @return {Promise<number[]>}
+ * Reads a file as binary.
+ *
+ * @param filePath Path to the file
+ * @param [options]
+ * @returns A promise resolving to an array of the file bytes.
  */
 async function readBinaryFile(
   filePath: string,
@@ -99,14 +99,11 @@ async function readBinaryFile(
 }
 
 /**
- * writes a text file
+ * Writes a text file.
  *
- * @param file
- * @param file.path path of the file
- * @param file.contents contents of the file
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @param file File configuration object
+ * @param [options]
+ * @returns
  */
 async function writeFile(
   file: FsTextFileOption,
@@ -133,10 +130,10 @@ async function writeFile(
 const CHUNK_SIZE = 65536
 
 /**
- * convert an Uint8Array to ascii string
+ * Convert an Uint8Array to ascii string.
  *
  * @param arr
- * @return ASCII string
+ * @returns An ASCII string.
  */
 function uint8ArrayToString(arr: Uint8Array): string {
   if (arr.length < CHUNK_SIZE) {
@@ -153,10 +150,10 @@ function uint8ArrayToString(arr: Uint8Array): string {
 }
 
 /**
- * convert an ArrayBuffer to base64 encoded string
+ * Convert an ArrayBuffer to base64 encoded string.
  *
  * @param buffer
- * @return base64 encoded string
+ * @returns A base64 encoded string.
  */
 function arrayBufferToBase64(buffer: ArrayBuffer): string {
   const str = uint8ArrayToString(new Uint8Array(buffer))
@@ -164,14 +161,11 @@ function arrayBufferToBase64(buffer: ArrayBuffer): string {
 }
 
 /**
- * writes a binary file
+ * Writes a binary file
  *
- * @param file
- * @param file.path path of the file
- * @param file.contents contents of the file
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @param file File configuration object
+ * @param [options]
+ * @returns
  */
 async function writeBinaryFile(
   file: FsBinaryFileOption,
@@ -196,13 +190,11 @@ async function writeBinaryFile(
 }
 
 /**
- * list directory files
+ * List directory files.
  *
- * @param dir path to the directory to read
- * @param [options] configuration object
- * @param [options.recursive] whether to list dirs recursively or not
- * @param [options.dir] base directory
- * @return
+ * @param dir Path to the directory to read
+ * @param [options] Configuration object
+ * @returns
  */
 async function readDir(
   dir: string,
@@ -219,15 +211,13 @@ async function readDir(
 }
 
 /**
- * Creates a directory
+ * Creates a directory.
  * If one of the path's parent components doesn't exist
- * and the `recursive` option isn't set to true, it will be rejected
+ * and the `recursive` option isn't set to true, it will be rejected.
  *
- * @param dir path to the directory to create
- * @param [options] configuration object
- * @param [options.recursive] whether to create the directory's parent components or not
- * @param [options.dir] base directory
- * @return
+ * @param dir Path to the directory to create
+ * @param [options] Configuration object
+ * @returns
  */
 async function createDir(
   dir: string,
@@ -244,14 +234,12 @@ async function createDir(
 }
 
 /**
- * Removes a directory
- * If the directory is not empty and the `recursive` option isn't set to true, it will be rejected
+ * Removes a directory.
+ * If the directory is not empty and the `recursive` option isn't set to true, it will be rejected.
  *
- * @param dir path to the directory to remove
- * @param [options] configuration object
- * @param [options.recursive] whether to remove all of the directory's content or not
- * @param [options.dir] base directory
- * @return
+ * @param dir Path to the directory to remove
+ * @param [options] Configuration object
+ * @returns
  */
 async function removeDir(
   dir: string,
@@ -268,13 +256,12 @@ async function removeDir(
 }
 
 /**
- * Copy file
+ * Copys a file to a destination.
  *
- * @param source
- * @param destination
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @param source A path of the file to copy
+ * @param destination A path for the destination file
+ * @param [options] Configuration object
+ * @returns
  */
 async function copyFile(
   source: string,
@@ -293,12 +280,11 @@ async function copyFile(
 }
 
 /**
- * Removes a file
+ * Removes a file.
  *
- * @param file path to the file to remove
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @param file Path to the file to remove
+ * @param [options] Configuration object
+ * @returns
  */
 async function removeFile(
   file: string,
@@ -317,11 +303,10 @@ async function removeFile(
 /**
  * Renames a file
  *
- * @param oldPath
- * @param newPath
- * @param [options] configuration object
- * @param [options.dir] base directory
- * @return
+ * @param oldPath A path of the file to rename
+ * @param newPath A path of the new file name
+ * @param [options] Configuration object
+ * @returns
  */
 async function renameFile(
   oldPath: string,
